✨ feat: 녹음 기능 모듈화 및 음성 분석 결과 표시 개선
